+++ src/settings/descriptors/line/lineDescriptor.ts
@@ -320,10 +320,6 @@ export class LineDescriptor extends BaseDescriptor {
         this.displayNameKey = "Visual_Line"
     }
 
-    public clearContainers(): void {
-        this.container.containerItems = [this.allLinesContainerItem];
-    }
-
     public populateContainer(dataPoint: LineDataPoint) {
         const { containerKey, containerName, selectionId, objects } = dataPoint
         const existingContainer = this.getCurrentSettings(containerKey)